requirements.txt
PyGithub>=1.59
requests>=2.31.0
python-dotenv>=1.0
flask>=3.0
flask-cors>=4.0
